Script para criar base de dados

+++ scripts/initializingProduct.sh
@@ -1,8 +1,16 @@
-# $1 parametro teste
+# $1 parametro DataBaseName
 echo "# Initializing $1 Product #"
-rm -rf vendor
-composer clear-cache
-composer update
-php artisan make:migrate
-php artisan db:seed
+if [ -z $1 ]; then
+	echo "Favor informar Base de Dados"
+else
+	echo "Atualizando Composer"
+	rm -rf vendor
+	composer clear-cache
+	composer update
+	echo "Criando Base de Dados $1"
+	php artisan mysql:createdb $1
+	echo "Criando seeders"
+	php artisan db:seed
+fi
+
 echo "# Done $1 Product #"
